Initial support for GigE machine vision cameras.
This commit enables initial support for Point Grey machine vision
cameras, and in general, for any camera supported through Project
Aravis.

These cameras are mainly geared towards high speed video streams, and do
not typically support very high exposure times. This driver attempts to
operate the camera in the most manual mode possible. Because most
cameras slave the programmable exposure time to the configured
framerate, which cannot be set lower than 1.0 FPS, out of the box these
cameras do not support longer exposure times, unless the frame rate is
manually disabled.

Unfortunately, this is a vendor-specific operation, here implemented for
the "BlackFly" family from Point Grey.
However, typically, it is not so difficult to set disable the frame
rate. Look in arvcamera.c (in Project Aravis) or BlackFly.cpp for clues
if you want to adapt this for your own camera.

Tested with kstars/EKOS and PHD2, with a BlackFly BFLY-PGE-31S4M camera.