Fix sequence of equal exposures causing frame corruption on DMx 21A04.AS
On Imaging Source DMx 21AU04.AS (199e:8101), multiple exposures with
same duration cause corrupted buffers to be notified to the indi driver.
One 1012-byte frame, then one 307200-byte frame, will be sent on every
capture after the first one. The 1012-byte frame matches the beginning
of the valid final frame, but cannot be used as is and often causes the
decoder to break after a few subsequent captures. Because the 1012-byte
frame arrives before exposure completion, the reported exposure time is
always shorter than the requested exposure time.

I first worked around this issue by checking the size of the frame based
on its width, height and depth, and could build a simple algorithm
dropping incorrect frames. But the workaround would only work with GREY
8/16 single-plane pixel formats.

However, unconditionally setting the exposure duration via
setINTControl's ioctl works around the problem. With DMx 21AU04.AS,
first frame is always 307200 bytes, and the resulting exposure time is
always greater than the requested exposure time.

Except for efficiency, it doesn't seem a problem to set the exposure
again for each frame. Apparently streaming is not affected by the issue.