Fixing infinite exposure loop when requesting duplicate exposure.
When requesting a new exposure while one is already running, V4L2_Driver
returns an error but doesn't stop V4L2_Base. As a consequence of the
error, INDI::CCD flags the end of the exposure, which causes V4L2_Driver
to ignore the frames V4L2_Base forwards forever.

Issue can be reproduced by starting a 10-second exposure in the "Main
Control" tab in Ekos, and clicking a second time the "Set" button.

To fix the infinite loop, V4L2_Driver aborts V4L2_Bases's exposure when
it receives a frame from V4L2_Base but PrimaryCCD isn't exposing. This
causes V4L2_Driver to reject exposures until V4L2_Base finishes the
current one (which will be discarded), at which point an exposure can be
restarted. However, the end-user sees an error icon and has no way to
know when the exposure finishes from the user interface.

To avoid losing the frame V4L2_Base is currently capturing, V4L2_Driver
returns success when asked to start a new exposure while one is already
running. As a consequence, the exposure is not started, but INDI::CCD
considers the driver is busy, and waits patiently with a busy icon for
the frame to arrive. The end-user is notified as KStars displays the new
frame, and can restart a new exposure if needed.