Letting auto exposure run if manual/auto cannot be controlled.
Some devices do not offer manual/auto exposure control (DMK31 for
instance, Raspberry NOIR probably). In order to trigger an exposure, the
driver must fall back to auto aperture mode and launch the exposure
nonetheless.

There are multiple possible situations in that case, depending on the
device, but bypassing manual/auto exposure control on a device providing
that feature actually switches the exposure time to the auto exposure
time currently configured.

This commit is thus an attempt to trigger that switch for devices not
providing manual/auto control. However, the actual behavior needs to be
validated by issue OPs as device vary greatly. There's limited chance
that the commit cause regressions, as the normal flow is not modified.