Fixed a bunch of issues that caused the NStep driver to not work
Added absolute capabilities.
Fixed an issue with the steppingMode reading the wrong values from the NStep.
Added funtionality to deal with steps > 999.
Fixed an issue that was caused by the fact that the NStep speed is faster as the numbers get lower.
Added SteppingPhase switch.
Fixed an issue where the client wasn't getting updated with the new position.