Changes to allow for FocuserInterface for Focuser #1, without renaming AbortSP, and to clean up some of the interface.
FocuserInterface does not allow for Focuser #2 in use, but does retain the old system that azwing created, so it can be used manually.

Kstars 2.9.4+ from git or the ubuntu ppa is *highly* recommended, as there appears to be an issue with multiple requests for exposures. See here for details: https://bugs.kde.org/show_bug.cgi?id=392974