OnStep: Extended GOTO error codes.  (#26)
* OnStep: TELESCOPE_SLEW_RATE will work properly now, relfects MaxSlewRateNP which was left in because it reports easily to the user the OnStep Value.

Also adds (initial, and as of yet unused code for :Gu#, should be a separate commit, but got started on two things, and doesn't harm anything)

* OnStep: Fix issue with Align being marked done prematurely

* OnStep: Fix other align command return issue

* OnStep: Fix tracking updates

* OnStep: Add support for reporting Guiding rates (GUIDE_RATE_NS and GUIDE_RATE_EW)

Also: Preparation for :Gu# support

* fixed format

* OnStep: Add Support for Refine Polar Alignment (:MP#) and instructions

* OnStep/LX200: Added Support for extended responses to :MS# errors

Current drivers should if there are responses of 3+, display simply the default of LOG_ERROR("Slew Failed"), changed to virtual in lx200telescope.cpp to allow overriding in OnStep. If an error code is not a number then it now returns -1. (Which will trigger the same default behavior.)