Fix spurious error in new Xagyl driver
This bug was introduced after being handled correctly in the middle of
my previous patch.

When issuing a (G)oto command, the device firmware will sometimes return
the message: "ERROR 7 - Magnet weak" after returning the successful
position response.

In this case, the error should be printed and ignored rather than
treated as a failure.