Rework Telescope Simulator
THis is a major change, the mount control now uses new Angle, Axis and
Alignment classes to handle the simulated mount.
The mount can be either afork mount equatorial mount or a GEM.
The GEM implements side of pier, the fork mount does not.
The AltAz option is not implemented.