Attempt to read the Pointing State from iOptron mounts
This is an experiment to use the :GEA# command to read the mount
axis values and use the dec axis to determine the pointing state.
The Pole Side cabability is only set if the GEA command returns a value.

This assumes that the Pointing State depends on the sign of the dec axis
value.