Implement Observed to J2000 catalogue conversion correctly
The Observed JNow to J2000 catalogue conversion did not remove
aberration or nutation and this was giving alignment errors of 20 to 40
arc seconds in the CCD and Guide Simulators.

A LibAstro class has been added which implements the conversions between
J2000 Catalogue and Observed JNow conversions allowing for precession,
nutation and aberration correctly in both directions.

This conversion appears to be implemented with the same problems in
other drivers so the conversions have been added to a new LibAstro
class.

The code continues to use the libnova code.