Update CCD and Guide simulators to improve image accuracy
The conversion of the JNow position to a J2000 catalogue position for
the GSC catalogue was  not done correctly, only the Precession was
removed, leaving aberration and nutation.

This adds libastro.cpp,.h files which contain converters to and from
J2000 catalogue and apparent positions.  When converting apparent to
J2000 the aberration, nutation and precession are all removed.  These
converters may be useful elsewhere, this isn't the only place
conversion from aparrent to J2000 is needed.