Changed the math include because ONE of the nine CI builds failed.
Typical Linux. No such thing as a standard. Just do anything you like
and leave if to the poor users to cope with the mess you leave behind.
And documentation consists of social media posts - if you are lucky.

It still builds, but then it did before so that says nothing.