Experimental update to Pier Side
This changes how pier side is determined:
If the hour angle is more than 2 hours from the meridian use hour angle.
If the declination pole angle is less then the difference between the
declination poile angle and the mount axis pole angle use the axis pole
angle
Otherwise return unknown.
This should be OK except when close to the meridian and with a
declination within a degree or two of the pole.