Change Pier Side update
Move the reading of the Pier Side from the mount to just after reading
the Ra and Dec so the values are will be synchronised.
Update the Pier Side just before the Ra and Dec.
The Pier side read from the mount doesn't match the Ha when the
declination is 90 so report the pier side as unknown when the
declination is close to 90 (>89.999).

This is to mitigate a problem where a user reported that the Ha limit
check failed during a pier flip slew.