Fix QHYCFW3 driver handshake
The QHYCFW3 driver tries to issue the VRS command immediately
after connection, but the filter wheel protocol will first
do a full turn or sometimes more than that, and report
the current filter position before that.

The process takes considerable time and always causes a timeout,
but even if the timeout is increased, the initial position
report will desync commands if not read.

This adds a read to expect that initial report with a longer
timeout that lets it do a full initialization under most
conditions.

Tested with a QHYCFW3-S