Update celestrondriver.cpp
The code tests if the degrees passed to it are > 0 if they are it sends East to the mount, if less than 0 it sends West, for location less than 1 degree East the code was wrongly sending West because the degrees are 0. My location is 000:07:24 East so I noticed the error. Checking for >= fixes the issue.