Script: Replace hard coded directory with variable
Until now, the developer-build script assumed a hard-coded directory
inside the RAM shared memory. A developer might not want to build the
final output into the /dev/shm directory.

With this commit, first get the desired build directory from
user. Check if it is in shared memory, i.e. /dev/shm/*, if so, check
if there is enought memory available. If not, abort. Else, check if
the directory exists. If not, create with its parents. Finally, create
a symbolic link to the correct destination.