Script: Update free space check/skip, fix -o bug
Now the user can skip checking for free space using the -f option. I
decided to add this function because /dev/shm can have a smaller value
than RAM itself, specially in a docker environment where one must feed
the 'docker run' with '--shm-size' option to increase the size based
on their needs.

Also, the script now checks if there is enough free space left in the
target build directory, no matter where it is located, e.g. /dev/shm
or not.

Lastly, if the '-o' option was followed by another option, e.g. '-f',
the script mistakenly took the following option as an argument for the
'-o' option. Now this bug is taken care of with checking the argument
fed to '-o'. If it starts with '-' then it is not a path. So, the
script will inform the user and exit.