Cleanup FocusLynx device types
Removed OC as it is hard coded into models that support it and should not be set in other controllers
Removed OE and OF. These are reserved and have not been used.
Changed name for FA and FB to note that they work for both the Optec and Starlight focusers. These are now the Generic QuickSynx / HSM device types.
Changed name for FD and FF to note that they work for both the Optec and Starlight focusers. These are now the Generic DirectSync / PDMS device types.
Hid the duplicate FA / FB device types. The new names should cover the types.
Added new in use device types OL (Sagitta 2) and RA (Unipolar motor)