Initial release of Alluna TCS2 Driver
The Telescope-Control-System TCS of ALLUNA Optics, has been designed
for the remote control of high-quality telescopes. Alluna Optics 
provides only ASCOM drivers and a Windows EXE program to control the 
TCS2 Device. The TCS2 can also be controlled by an optional handheld 
device. 

The indi driver has been developed by me, Peter Englmaier, and I am 
not affiliated with Alluna Optics, but own a small Alluna Telescope 
and want to use it with Ekos and other indi software.

With TCS several telescope functions are controlled and monitored:
- Air conditioning of the optical tube and the main mirror
  i.e., primary/secondary mirror heating and ventilation
- Temperature of ambient (focuser), primary and secondary mirror
- Focusing of the optics
- Control of the optionally available CCD Rotator
- Control of the optionally available robotic Dust Cover
- Setting configuration parameters

As I do not have the rotator option, I have not (yet) implemented that 
part. Also some less useful functions are still missing. 

But it is possible, to use the focuser with ekos, control the dust 
cover, and the climate settings.