Fix version reporting on older Celestron NexStar mounts (#1988)
Celestron NexStar mount hand controllers have been introduced with a
number of firmware versions over the years.  Starting with the 4.xx
NexStar controllers, the minor version is a 2-digit value (currently
4.22).  The older hand controllers (while sporting fewer features) still
will talk to INDI since the basic serial communication protocol has not
changed.

For these older controllers (v1.2, v1.6, and v2.2), the minor version is
a single-digit value.  The current code in CelestronDriver::get_version
forces a two-digit minor version, so, for instance, the v2.2 controller
is recorded as "2.02".  When the version is checked against device
capabilities, 2.02 < 2.2, so it is incorrectly rejected.

This commit checks the major version of the hand controller, and uses
a 2-digit minor version for values larger than 2, and a 1-digit minor
version for earlier models.

	modified:   drivers/telescope/celestrondriver.cpp