Astrotrac 360 Driver Motion Control Issue (#2052)
Moving the mount N/S or E/W via the motion buttons in the indi "Motion
Control" tab can inadvertently change the tracking state, even if the
scope is idle, because the MoveNS and MoveWE functions always apply the
custom tracking rates, which are set to sidereal by default, when motion
is stopped. Fixed this issue by calling SetTrackEnabled when motion is
stopped to reset tracking. Tested with my Astrotrac 360 mount.